requirements.txt
numpy>=1.26
pandas>=2.2
PyYAML>=6.0
networkx>=3.2
matplotlib>=3.8
